% Generated by roxygen2 (4.0.2): do not edit by hand
\name{brownie.extract.file.names}
\alias{brownie.extract.file.names}
\title{Extracts file names from the meta files. Needs recognized structure from meta.json}
\usage{
brownie.extract.file.names(exp.meta)
}
\description{
Extracts file names from the meta files. Needs recognized structure from meta.json
}

