% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.sync}
\alias{check.sync}
\title{Checks synchronization between EDF and DAT files}
\usage{
check.sync(edffile, eegfile)
}
\arguments{
\item{edffile}{an .edf file name without an extesion}

\item{eegfile}{a .dat file name without an extesion}
}
\value{
a table with all sync points and their timestamps for both files.
Also saves the same table to asc file with a corresponding name.
}
\description{
Computes differenece between every two synchronization points in the beggining and in the end of two files
}

