% Generated by roxygen2 (4.0.2): do not edit by hand
\name{classify}
\alias{classify}
\title{Prepares eeg and eyetracking data and performs claassification}
\usage{
classify(exp_type, exp_name, classifier = NULL)
}
\arguments{
\item{exp_type}{name of the experiment's serie (e.g. "e2")}

\item{exp_name}{name of the concrete experiment (e.g. "01" if your experiment's full name is "e201s001")}
}
\description{
Prepares eeg and eyetracking data and performs claassification
}
\examples{
#First need to set base path for the Brownie package
#that contains a folder with data
set.brownie.root("d:\\\\")
}

