% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decimate.with.na}
\alias{decimate.with.na}
\title{Decimates a vector, that contains NAs}
\usage{
decimate.with.na(vect, q, n = if (ftype == "iir") 8 else 30, ftype = "iir")
}
\arguments{
\item{q}{integer factor to downsample by.}

\item{n}{filter order used in the downsampling.}

\item{ftype}{filter type, "iir" or "fir"}

\item{x}{signal to be decimated, may contain NAs.}
}
\description{
Based on decimate function from the "signal" package.
By default, an order 8 Chebyshev type I filter is used or a 30-point FIR filter if ftype is 'fir'. Note that q must be an integer for this rate change method.
Makes use of the filtfilt function with all its limitations.
}
\examples{
decimate.with.na(x, q)
}
\author{
Nuzhdin Urii, Anastasia Fedorova
}

