% Generated by roxygen2 (4.0.2): do not edit by hand
\name{many.files.any.threshold}
\alias{many.files.any.threshold}
\title{Ascii for eeglab with comparison for many files simultaniously}
\usage{
many.files.any.threshold(actualFixDur, lowerThresh, higherThresh, files)
}
\arguments{
\item{actualFixDur}{fixation threshold that was used in experiment}

\item{lowerThresh}{lower threshold to use in comparison}

\item{higherThresh}{higher threshold to use in experiment}

\item{files}{list of files to be processed}
}
\description{
Creates ascii file for eeglab with fixation's starts
that were proceessed in offline simulation and possibly compared with two fixation thresholds
also puts all saccades ends in ascii file
}
\details{
If you need to compare only one file use arb.thresh function
}
\examples{
many.files.any.threshold(800, 500, 1000, j.files800)

#this will take all files where threshold was 800 and compare each of them
#as if the threshold was between 500 and 1000
}

