% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decimate.eye}
\alias{decimate.eye}
\title{Decimates samples from EDF file}
\usage{
decimate.eye(filename, q)
}
\arguments{
\item{filename}{- name of a file with extension}

\item{q}{- integer factor to downsample by. E.g. if your file's sampling rate is 1000 Hz and you want to downsample it to 500 Hz, q should equal "2".}
}
\value{
Writes a table to a working derictory in ascii format compatible with EEGlab
}
\description{
Decimates samples from EDF file
}
\examples{
decimate.eye('23566333.edf', 2)
}

